Refresh Site Info on Cache Clear
Grabs site info from TCB server when cache is cleared on client. The
method of achieving this is crude, as the TCB server caches responses to
Drupal clients but not to other clients. The cache clearing for some
reason on other clients works (using Drupal's dynamic caching) but for
some reason requests from another Drupal site hit Drupal's lower level
of caching (Internal Caching) which doesn't seem to clear. The way
around this for now is to add a unix timestamp at the end of each
request from the TCB client, but a more elegant solution should be found
for the future.