Dependency Maintenance and Requires
Added a requirement in the .info.yml file to specify the dependency
modules must be installed before using tcb auth client (completing #6).
Also, did some maintenance on the TCBUserAuthenticator and
TCBUserManager classes to make their code in-line with the changes in
the latest versions of social auth and social api modules. This
completes #8.