Change Form Label and Make Admin Default
Implemented the features requested in issues #9 and #13. On the config
form, the text box label was changed to clarify that a domain is needed
as a value, not a URL. Help text was provided to give an example of what
is required. Additionally, when a new user is created they are
automatically made an administrator by default.