Assign Custom Role on User Creation
Users, if defined on the TCB Server, now will be created with the role
assigned to on the TCB Server--if the role exists locally (on the client
site). If the role doesn't exist locally, or there is no user defined
for the user's email, then the user is created with the site default role.