Bug Fix on Permissions Update
Fixed the bug described in #17 where permissions on TCB roles would be
revoked and not restored, regardless of whether the user actually
changed the role permissions. Since TCB role permissions should always
be the same as those configured on the server, I have made it so that
permissions on TCB roles are always set back to what they're configured
to.